fix: move utils inside
