+++ packages/components/src/components/dialog/dialog.component.ts
@@ -6,7 +6,6 @@ import { Component } from '../../models';
 import { FocusTrapMixin } from '../../utils/mixins/FocusTrapMixin';
 import { DEFAULTS } from './dialog.constants';
 import type { DialogRole, DialogSize, DialogVariant } from './dialog.types';
-import { DialogUtils } from './dialog.utils';
 import { TYPE, VALID_TEXT_TAGS } from '../text/text.constants';
 import { DialogEventManager } from './dialog.events';
 import { BUTTON_VARIANTS, ICON_BUTTON_SIZES } from '../button/button.constants';
@@ -153,28 +152,64 @@ class Dialog extends FocusTrapMixin(CardAndDialogFooterMixin(Component)) {
   override role: DialogRole = DEFAULTS.ROLE;
 
   /** @internal */
-  private triggerElement: HTMLElement | null = null;
+  protected triggerElement: HTMLElement | null = null;
 
   /** @internal */
-  private utils: DialogUtils;
+  protected backdropElement: HTMLElement | null = null;
 
   /** @internal */
-  private backdropElement: HTMLElement | null = null;
-
-  /** @internal */
-  private lastActiveElement: HTMLElement | null = null;
+  protected lastActiveElement: HTMLElement | null = null;
 
   constructor() {
     super();
-    /** @internal */
-    this.utils = new DialogUtils(this);
     document.addEventListener('keydown', this.onEscapeKeydown);
+  }
+
+  /**
+   * Sets up the aria attributes for the dialog based on the header text and aria attributes.
+   * If no header text or aria attributes are provided, it will use the triggerId if available.
+   * @internal
+   */
+  private setupAriaAttributes() {
+    if (this.headerText && !this.ariaLabel && !this.ariaLabelledby) {
+      this.setAttribute('aria-labelledby', this.headerText);
+    } else if (!this.headerText && !this.ariaLabel && !this.ariaLabelledby) {
+      if (this.triggerId) {
+        this.setAttribute('aria-labelledby', this.triggerId);
+      }
+    }
+  }
+
+  /**
+   * Creates a backdrop element for the dialog.
+   * The backdrop is a full-screen overlay that appears behind the dialog when it is open.
+   * It prevents interaction with the rest of the application while the dialog is open.
+   * @internal
+   */
+  private createBackdrop() {
+    const backdrop = document.createElement('div');
+    backdrop.classList.add('dialog-backdrop');
+    const styleElement = document.createElement('style');
+    styleElement.textContent = `
+      .dialog-backdrop {
+        position: fixed;
+        top: 0;
+        left: 0;
+        width: 100%;
+        height: 100%;
+        background: var(--mds-color-theme-common-overlays-secondary-normal);
+        z-index: ${this.zIndex - 1};
+      }
+    `;
+    backdrop.appendChild(styleElement);
+    this.parentElement?.appendChild(backdrop);
+    this.backdropElement = backdrop;
   }
 
   protected override async firstUpdated(changedProperties: PropertyValues) {
     super.firstUpdated(changedProperties);
     this.setupTriggerListener();
-    this.utils.setupAriaAttributes();
+    this.setupAriaAttributes();
     this.style.zIndex = `${this.zIndex}`;
     DialogEventManager.onCreatedDialog(this);
 
@@ -229,7 +264,7 @@ class Dialog extends FocusTrapMixin(CardAndDialogFooterMixin(Component)) {
       changedProperties.has('aria-label')
       || changedProperties.has('aria-labelledby')
     ) {
-      this.utils.setupAriaAttributes();
+      this.setupAriaAttributes();
     }
   }
 
@@ -265,7 +300,7 @@ class Dialog extends FocusTrapMixin(CardAndDialogFooterMixin(Component)) {
 
       this.enabledFocusTrap = true;
       this.enabledPreventScroll = true;
-      this.utils.createBackdrop();
+      this.createBackdrop();
 
       await this.handleCreateDialogFirstUpdate();
